+++ src/main/java/com/structurizr/dsl/IdentifierScope.java
@@ -2,7 +2,7 @@ package com.structurizr.dsl;
 
 public enum IdentifierScope {
 
-    Global,
-    Local,
+    Flat,
+    Hierarchical,
 
 }