+++ src/main/java/com/structurizr/dsl/IdentifierScopeParser.java
@@ -2,12 +2,12 @@ package com.structurizr.dsl;
 
 final class IdentifierScopeParser extends AbstractParser {
 
-    private static final String GRAMMAR = "!identifiers <global|local>";
+    private static final String GRAMMAR = "!identifiers <flat|hierarchical>";
 
     private static final int MODE_INDEX = 1;
 
     IdentifierScope parse(DslContext context, Tokens tokens) {
-        // !identifiers <global|local>
+        // !identifiers <flat|hierarchical>
 
         if (tokens.hasMoreThan(MODE_INDEX)) {
             throw new RuntimeException("Too many tokens, expected: " + GRAMMAR);
@@ -18,10 +18,10 @@ final class IdentifierScopeParser extends AbstractParser {
         }
 
         String name = tokens.get(MODE_INDEX);
-        if ("global".equalsIgnoreCase(name)) {
-            return IdentifierScope.Global;
-        } else if ("local".equalsIgnoreCase(name)) {
-            return IdentifierScope.Local;
+        if ("flat".equalsIgnoreCase(name)) {
+            return IdentifierScope.Flat;
+        } else if ("hierarchical".equalsIgnoreCase(name)) {
+            return IdentifierScope.Hierarchical;
         } else {
             throw new RuntimeException("Expected: " + GRAMMAR);
         }